Move Aggregator and implementation to aggregators package
